Define "noto" construct.
We defined the "not" operator in logic programming as "noto" in stable-Kanren, since the keyword "not" is a reserved keyword in Scheme.

The new construct "noto" only does one thing, increasing the negation counter by one.

So negation as failure means that the goal function returns a new substitution if we are failed to prove a goal. In stable-Kanren goal function returns a new substitution only on a success proof. To reuse this mechanic, we are letting the goal function prove the complement of the goal. If the goal function successfully proved the complement goal, it failed to prove the original goal.