Add disunification.
When n is an even number, the unification worked as usual. If the unification succeeds, the substitution expands; otherwise, it fails. When n is an odd number, the unification worked as disunification. If the unification succeeds, it fails; otherwise, we didn't expand the substitution. Ideally, under the disunification, we can maintain a list of values the variable can't take. MiniKanren had an extension to inequality constraint, we will leave this as future work to the constraint support.