Add transformers to complement the rule.
Add new language construct 'defineo', so we can implicitly perform the rule translation.

During the execution, the goal function picks the corresponding rule set based on the value of the negation counter.
            n >= 0 and even use original rules
            n >= 0 and odd use complement rules

Add a new transformer, "complement" to transform the original rule to the complement form.

A helper transformer, "conde-t" turns the conjunction of goals into disjunctions of goals.