Add call frame stack to the goal function.
We add a call frame stack to the goal function's parameter so that the goal function will use the calling history to avoid a loop. If the current call frame is shown on the stack, we handle the loop.

Any variables passed to and returned by the goal function can be treated as the accumulated result. The variable only passes to the goal function will be consumed and helped with runtime computation only.

Hence our newly introduced negation counter is served the later purpose. All goal functions, including the special "conde" and "fresh" are all connected through "bind*"/"bind", bind* needs to hold the counter value at this goal function and pass it down to "bind" on each sub-goal function. The same changes applied to the other special goal functions like ifa, condu, ifu, and project.

We are showing how to use this call frame stack in the next part.