Add transformers to complement fresh(exist). Part1
Our previous implementation of complement and conde-t applies the DeMorgan rule on the disjunction of a set of rules in conjunctive normal form (CNF). The transformation only works for propositional logic. To evolve the transformer to support predicate logic, we need to handle the rule with a "fresh" (existential quantifier) operator. The existential quantifier introduces a new temporary variable in the body of the rule. Therefore, we can't simply apply the DeMorgan law as we did for the propositional program. The temporary variable may or may not bind to a value during the evaluation process; a predicate can serve as a generator to assign a value to the temporary variable or as a checker to check the value of the temporary variable still holds.

For example,
edge(a, b), edge(b, c), edge(a, d).

reachable(X, Y) :- edge(X, Y).
reachable(X, Y) :- edge(X, Z), reachable(Z, Y).

reducible(X) :- reachable(X, Y), not reachable(Y, X).

The former predicates edge(X, Z), and reachable(X, Y) in rule "reachable" and "reducible" respectively, are served as the generator. The latter predicates edge(Z, Y) and reachable(Y, X) are served as the checker.

The complement of the generator is simple. "There is an edge(X, Z)" to "There is no edge(X, Z).", but the complement of the checker needs to include the mutually exclusive of the first part. "If there is an edge(X, Z), then Z to Y is not reachable."

The complement rule of "reachable" is saying that "There is not exist such Z, that edge(X, Z), reachable(Z, Y) is true." The equivalent statement is "For all values of Z, that there is no edge(X, Z), or if there is an edge(X, Z), there is no reachable(Z, Y)."

So the complement of fresh should be a disjunction of the negation to each sub-goal in conjunction with all sub-goals before the current one.