Fix bug: transform the rule without "fresh".
During the testing, we found out that the program was trapped in an infinite loop for the query "(run* (q) (reducible q) )".

For predicate logic, even for the rule without "fresh" can't directly apply DeMorgan law. The two variables in one predicate are related to each other.

Take the "reduce(x, y)" as an example,

(defineo (reduce x y)
  (conde
    [(== x 'a) (== y 'b)]
    [(== x 'b) (== y 'c)]
))

The negated form is
x /= 'a, x = 'a; y /='b, x /= 'b, x = 'b; y /= 'c.
instead of
x /= 'a, y /= 'b, x /= 'b, y /= 'c.

Therefore, we need to perform the same transformation to the predicate without "fresh" in it.