Add transformers to complement fresh(exist). Part2
Once we finished the transformation to the original rule, all existential quantifiers over the variables in the rule turned into the universal quantifier. Hence, we need to convert the search stream from finding one answer (existential) to finding all answers (universal). For those inference variables, the finding of all answers is taken care of already due to the implicit universal quantifier in the rule's head.

We modify "fresh-t" to find out the generator to the intermediate variable, then we get the domain of the variable by assuming the domain space is finite using the "domain-values" macro. At last, we iterate through the domain and create a conjunction search stream over all possible values through the "forall" macro.