Add nqueens problem top-down solution.
To solve the nqueens problem in stratified negation, the programmer still has to direct the search process and backtrack on the conflicted queens in the "attack" predicate. Then the negation over "attack" turns the constraints from DNF into CNF.

The "q(X, Y)" is served as the "functional term" in the program, the two loop rules involving predicates "q(X, Y)" and "negq(X, Y)" and the "q(X, Y)" in the body of "pickqueen" do not impact the solving process. Removing those from the example, we can still get the correct answer and the performance remains the same.

The underlying miniKanren is a fully relational programming system that lacks fast numeric computation support. All numbers are defined as reversed binary and the arithmetic operations are shifting one number on the number axis. There are some impure operators in Prolog/s(ASP), such as "is", ">", and "-", so we also need to provide the counterpart in stratKanren. So, the calculation can fully utilize the modern CPU.