Add partial result along with subsititution.
For a normal program, it could produce multiple answer sets, and each answer set may have different atoms to be true. Therefore, we are saving the partial result we have obtained to avoid the future-solving process of flipping the outcome. The two-person game is an excellent example of this case: "a" and "b" can be true but not simultaneously.