Add partial result to the goal function.
Before the execution, check if we have computed the partial result. If the partial result has the same parity as n, returning true. Otherwise, returning false. After the execution, we memorize the partial result. At this point, all arguments have a substitution. So the partial result will record the predicate with actual values. We can't use the signature for ext-p here, as the signature was obtained before the execution of the goal function.