Add global constraint checking.
For a normal program, we need to check all the rules, especially those with negation literals inside. So that we make sure we find a stable model. For each rule, we need to check all possible values of head variables. So we use the reduct program to get all value sets of a given goal with unbounded variables. The reduct program is a special program without any negations.

[ToDo] Performance optimization, current `check-all-rules` computes all future answers, but we only need to verify one. Also, stratified programs don't need such checking, we are still deciding which way we want to choose.